NEW:  > VideoHandler (NativeModule) - this is used to handle creation of  video thumbnails once a video is successfully cached onto the system.
UPDATED:
 > FileManager (utilities/file_manager) - addition of the
 FileManager.fetchThumbnail to interface with the NativeModule that
 allows fetching of Thumbnails.
 > Addition of video support on the ReportDetails route, done on the
 file manager module and also on the ReportDetails screen.
 > Video carousel also supports thumbnails now on its individual cards
 > API -made a quick fix on media links addition to reports.
 >RemoteFetchWorkerThread (NativeModule) - intergrates with the
 VideoHandler thread.

+++ android/app/src/main/java/com/digital_matatus/utilities/RemoteFetchWorkerThread.java
@@ -12,6 +12,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.InputStream;
 import java.io.FileOutputStream;
+import java.net.HttpURLConnection;
 import java.net.URL;
 
 public class RemoteFetchWorkerThread implements Runnable {
@@ -22,6 +23,8 @@ public class RemoteFetchWorkerThread implements Runnable {
   // this will be used to resolve the content from the remote API
   private final File saveableFile;
   private final ContentResolver contentResolver;
+
+  private File THUMB_DIRECTORY = null; // not null when it is a video file
 
   /**
    * This is the Runnable class that will be used to fetch the file from the remote URI and store
@@ -42,40 +45,79 @@ public class RemoteFetchWorkerThread implements Runnable {
     this.saveableFile = saveableFile;
   }
 
+  public RemoteFetchWorkerThread(
+      URL remoteUri,
+      ReactApplicationContext reactApplicationContext,
+      File saveableFile,
+      File THUMB_DIRECTORY
+  ) {
+    this.remoteUri = remoteUri;
+    this.reactApplicationContext = reactApplicationContext;
+    this.contentResolver = reactApplicationContext.getContentResolver();
+    this.saveableFile = saveableFile;
+    this.THUMB_DIRECTORY = THUMB_DIRECTORY;
+  }
+
+  public RemoteFetchWorkerThread(File saveableFile) {
+    this.remoteUri = null;
+    this.reactApplicationContext = null;
+    this.contentResolver = null;
+    this.saveableFile = saveableFile;
+  }
+
   @Override
   public void run() {
 
-    try {
-      InputStream inputStream = remoteUri.openStream();
-      FileOutputStream fileOutputStream = new FileOutputStream(saveableFile);
+    if(this.remoteUri == null) // file is already created
+      this.emitEvent();
+    else {
+      try {
+        HttpURLConnection httpURLConnection = (HttpURLConnection) remoteUri.openConnection();
+        httpURLConnection.connect();
+        InputStream inputStream = httpURLConnection.getInputStream();
+        FileOutputStream fileOutputStream = new FileOutputStream(saveableFile);
 
-      int data = inputStream.read();
+        int data = inputStream.read();
 
-      while(data != -1) {
-        fileOutputStream.write(data);
-        inputStream.read();
+        while(data != -1) {
+          fileOutputStream.write(data);
 
-        data = inputStream.read();
+          data = inputStream.read();
+        }
+
+        inputStream.close();
+        fileOutputStream.flush();
+        fileOutputStream.close();
+        httpURLConnection.disconnect();
+
+        this.emitEvent();
+
+        // this block is only for the creation of thumbnails in the event that this is a video
+        // file that is being saved to cache.
+        if(THUMB_DIRECTORY != null) {
+          VideoHandler videoHandler = new VideoHandler(saveableFile, THUMB_DIRECTORY, reactApplicationContext);
+          Thread videoHandlerThread = new Thread(videoHandler);
+          videoHandlerThread.start();
+        }
+
+      } catch(Exception e) {
       }
-
-      inputStream.close();
-      fileOutputStream.flush();
-      fileOutputStream.close();
-
-      String cacheUri = "file://" + saveableFile.getName();
-
-      WritableMap params = Arguments.createMap();
-      params.putString("uri", cacheUri);
-
-      // emit the File.getName() event to help the unique listener get its unique
-      // listener when it's complete moving the data.
-      reactApplicationContext
-          .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
-          .emit(saveableFile.getName(), params);
-
-    } catch(Exception e) {
     }
 
+
+  }
+
+  private void emitEvent() {
+    String cacheUri = "file://" + saveableFile.getName();
+
+    WritableMap params = Arguments.createMap();
+    params.putString("uri", cacheUri);
+
+    // emit the File.getName() event to help the unique listener get its unique
+    // listener when it's complete moving the data.
+    reactApplicationContext
+        .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
+        .emit(saveableFile.getName(), params);
   }
 
 }
